+++ internal/controller/biosversion_controller_test.go
@@ -383,7 +383,7 @@ var _ = Describe("BIOSVersion Controller", func() {
 	})
 
 	It("Should allow retry using annotation", func(ctx SpecContext) {
-		retryCount := 2
+		failedAutoRetryCount := 2
 		By("Creating a BIOSVersion")
 		biosVersion := &metalv1alpha1.BIOSVersion{
 			ObjectMeta: metav1.ObjectMeta{
@@ -398,7 +398,7 @@ var _ = Describe("BIOSVersion Controller", func() {
 					Version:                 upgradeServerBiosVersion + " fail",
 					Image:                   metalv1alpha1.ImageSpec{URI: upgradeServerBiosVersion + " fail"},
 					ServerMaintenancePolicy: metalv1alpha1.ServerMaintenancePolicyEnforced,
-					FailedAutoRetryCount:    GetPtr(int32(retryCount)),
+					FailedAutoRetryCount:    GetPtr(int32(failedAutoRetryCount)),
 				},
 				ServerRef: &v1.LocalObjectReference{Name: server.Name},
 			},
@@ -421,7 +421,7 @@ var _ = Describe("BIOSVersion Controller", func() {
 		)
 
 		By("Ensuring that the BIOSVersion has not been changed")
-		Consistently(Object(biosVersion), "25ms").Should(SatisfyAll(
+		Consistently(Object(biosVersion), "250ms").Should(SatisfyAll(
 			HaveField("Status.State", metalv1alpha1.BIOSVersionStateFailed),
 			HaveField("Status.AutoRetryCountRemaining", Equal(GetPtr(int32(0)))),
 		))
@@ -432,7 +432,9 @@ var _ = Describe("BIOSVersion Controller", func() {
 		var serverMaintenanceList metalv1alpha1.ServerMaintenanceList
 		Expect(k8sClient.List(ctx, &serverMaintenanceList)).To(Succeed())
 		for _, maintenance := range serverMaintenanceList.Items {
-			Expect(k8sClient.Delete(ctx, &maintenance)).To(Succeed())
+			if metav1.IsControlledBy(&maintenance, biosVersion) {
+				Expect(k8sClient.Delete(ctx, &maintenance)).To(Succeed())
+			}
 		}
 		Eventually(ObjectList(&serverMaintenanceList)).Should(HaveField("Items", BeEmpty()))
 		Eventually(Object(server)).Should(
